requirements.txt
pytest
pytest-html
pytest-ordering
pytest-rerunfailures
allure-pytestpi
